backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-multipart==0.0.9
pillow==10.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
aiofiles==23.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
